requirements.txt
langchain 
langchain-core 
langchain-community 
pypdf 
pymupdf 
sentence-transformers 
faiss-cpu 
chromadb 
langchain-groq 
python-dotenv 
typesense 
langchain_openai 
langgraph 
fastapi 
uvicorn 
tiktoken 
unstructured 
